* FIX

See merge request cinogamaproject/woolang!421

+++ src/wo_api_impl.cpp
@@ -3082,14 +3082,14 @@ WO_API void wo_stack_value_set(wo_stack_value sv, wo_vm vm, wo_value val)
 {
     auto* vmbase = WO_VM(vm);
 
-    wo_assert(sv < vmbase->stack_mem_begin - vmbase->sp);
+    wo_assert(sv < (size_t)(vmbase->stack_mem_begin - vmbase->sp));
     wo_set_val(CS_VAL(vmbase->stack_mem_begin - sv), val);
 }
 WO_API void wo_stack_value_get(wo_value outval, wo_stack_value sv, wo_vm vm)
 {
     auto* vmbase = WO_VM(vm);
 
-    wo_assert(sv < vmbase->stack_mem_begin - vmbase->sp);
+    wo_assert(sv < (size_t)(vmbase->stack_mem_begin - vmbase->sp));
     wo_set_val(outval, CS_VAL(vmbase->stack_mem_begin - sv));
 }
